Docs: Update the demo code for Loading Button
**Problem:**
According to the current implementation of Blazor(https://github.com/dotnet/aspnetcore/issues/16213#issuecomment-381124723), it runs in single-thread mode. Thus the call to `System.Threading.Thread.Sleep` will block the UI thread as well. It means when clicking the loading button, the UI freezes. 
**Changes:**
Use `Task.Delay` which is a kind of built-in timer from dotnet. It won't block the UI thread.

*If we need to make more changes to make it happen in demo page, please let me know.*