feat(module: icon): icon role attribute change now  assignable, does not change default behavior (#3370)
when a screen reader parses an icon, it sees the role attribute has a value of 'img'. in the case where we don't want the screen reader to read the icon, the role can be assigned a blank value and the role attribute will not be added. instead, the aria-hidden attribute will be added with a value of 'true'

no breaking changes