Enhance Row component alignment options
Updated the `Row` component to set a default value for the `Align` property to `RowAlign.Top`. Added a new alignment option `Stretch` to the `RowAlign` enum. Modified class mapper logic to include `Stretch` alignment and added `SpaceEvenly` to the `RowJustify` enum. Updated the `SpaceAlign` enum to set a default value for `Start` to 1 for improved clarity and usability.