feat(module: upload): supports paste upload (#4650)
* feat(module: upload): supports paste upload

* Add Trigger property to Upload component

Introduces a new `Trigger` property in the `Upload` component of `Upload.razor.cs`. This property, marked with the `[PublicApi("1.5.0")]` attribute, allows users to specify the upload trigger mode, defaulting to `UploadTrigger.Click`. This enhancement improves the component's functionality and usability.

* Refactor UploadTests to ensure consistent input element rendering

Updated the UploadTests.razor file to maintain consistent rendering of input elements within the upload component. This change enhances the clarity and reliability of the tests by ensuring that the input elements are consistently defined across different upload scenarios.

* feat(module: upload): support pasting files from both input and textarea